requirements.txt
audioread==3.0.1
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==5.2.1
ffmpeg-python==0.2.0
Flask==3.0.3
flask-cors==5.0.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
msgpack==1.1.0
numba==0.61.2
numpy==2.2.5
packaging==25.0
platformdirs==4.3.7
pooch==1.8.2
pycparser==2.22
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
soundfile==0.13.1
soxr==0.5.0.post1
threadpoolctl==3.6.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
yt-dlp==2024.10.22
